Refactor code and add directory and repository support to _ask_doc function
